project structure refatoring
