--- Oxide.Ext.RustApi/Exceptions/ApiCommandNotFoundException.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace Oxide.Ext.RustApi.Exceptions
-{
-    /// <summary>
-    /// Route exception.
-    /// </summary>
-    internal class ApiCommandNotFoundException : Exception
-    {
-        public ApiCommandNotFoundException(string message) : base(message)
-        {
-            
-        }
-    }
-}
