+++ Oxide.Ext.RustApi/Plugins/RustApiPlugin.cs
@@ -1,0 +1,7 @@
+﻿namespace Oxide.Ext.RustApi.Plugins
+{
+    public class RustApiPlugin
+    {
+        
+    }
+}
